Added model + changed user interface
-Added actor subsearch model object
*TODO: Should try to abstract the data rather than returning the JSON
representation
*NOTE: Temporarily have made the query functions static, but should be
an instance method later
*TODO: Maybe create a base class for making get and post requests which
the other types of models
would feed into and parse the relevant results
-Modified actor subsearch to use the actor subsearch model object i.e.
populates results from tmdb
*TODO: May want to perform a second query immediately after if the
search has more than 1 page
such that there is no delay when the user scrolls to the end of the
first page results
or perhaps a "Load more results" button at the bottom
*TODO: Would be helpful to display more than just the names of the
actors or have some means of displaying
more detailed information, added an item to the context menu for now
that does nothing as a possibility
*TODO: Would like to refine selection... dialogs might disrupt flow but
not validating risks an accidental selection.
Was planning on using contextual action bars, but it requires a lot of
overriding of system behavior; may be an easier way.
*TODO: Still needs to pass data back to original actor search activity